# Conflicts:
#	app/src/main/java/com/example/yourschedule/DECORATOR/TextSizeDecorator.java
#	app/src/main/java/com/example/yourschedule/FRAGMENT/Calandar.java
#	app/src/main/res/layout/fragment_calandar.xml

+++ app/src/main/java/com/example/yourschedule/Formatter/TempValueFormatter.java
@@ -14,7 +14,7 @@ public class TempValueFormatter extends ValueFormatter {
     @Override
     public String getFormattedValue(float value) {
 
-        return value+"℃";
+        return (int)value+"℃";
 
     }
